fix bug parsing large integers that cannot fit in 32 bits

+++ include/JsonBox/Value.h
@@ -109,7 +109,7 @@ namespace JsonBox {
 		 * Constructs the value from an integer.
 		 * @param newInt Integer used as the value.
 		 */
-		Value(int newInt);
+		Value(int64_t newInt);
 
 		/**
 		 * Constructs the value from a double.
@@ -173,7 +173,7 @@ namespace JsonBox {
 		 * @param src Integer to copy.
 		 * @return Reference to the modified value.
 		 */
-		Value &operator=(int src);
+		Value &operator=(int64_t src);
 
 		/**
 		 * Assignation operator overload.
@@ -387,7 +387,7 @@ namespace JsonBox {
 		 * @return Value's integer value, or 0 if the value doesn't contain a
 		 * numeric value.
 		 */
-		int getInteger() const;
+		int64_t getInteger() const;
 
 		/**
 		 * Tries to get the value's integer value.
@@ -396,14 +396,14 @@ namespace JsonBox {
 		 * @return Value's integer value, or defaultValue if the value doesn't
 		 * contain a numeric value.
 		 */
-		int tryGetInteger(int defaultValue) const;
+		int64_t tryGetInteger(int64_t defaultValue) const;
 
 		/**
 		 * Sets the value as an integer.
 		 * @param newInt New integer value that the Value will contain. The
 		 * value's type is changed if necessary to contain the integer.
 		 */
-		void setInteger(int newInt);
+		void setInteger(int64_t newInt);
 
 		/**
 		 * Gets the value's double value.
@@ -557,7 +557,7 @@ namespace JsonBox {
 		 */
 		union ValueDataPointer {
 			std::string *stringValue;
-			int *intValue;
+			int64_t *intValue;
 			double *doubleValue;
 			Object *objectValue;
 			Array *arrayValue;
@@ -576,9 +576,9 @@ namespace JsonBox {
 
 			/**
 			 * Parameterized constructor.
-			 * @param newIntValue Pointer to set to the int pointer.
+			 * @param newIntValue Pointer to set to the int64_t pointer.
 			 */
-			ValueDataPointer(int *newIntValue);
+			ValueDataPointer(int64_t *newIntValue);
 
 			/**
 			 * Parameterized constructor.
